+++ test/db/isoexpress/mra_capacity_bidoffer_test.dart
@@ -2,13 +2,14 @@ library test.db.isoexpress.mra_capacity_bidoffer_test;
 
 import 'package:dotenv/dotenv.dart' as dotenv;
 import 'package:elec_server/src/db/isoexpress/mra_capacity_bidoffer.dart';
+import 'package:elec_server/src/db/lib_prod_archives.dart';
 import 'package:elec_server/src/db/lib_prod_dbs.dart';
 import 'package:test/test.dart';
 import 'package:timezone/data/latest.dart';
 import 'package:date/date.dart';
 
 Future<void> tests(String rootUrl) async {
-  var archive = MraCapacityBidOfferArchive();
+  var archive = getIsoneMraBidOfferArchive();
   group('MRA BidOffer archive tests:', () {
     test('read file for 2024-01', () async {
       var file = archive.getFilename(Month.utc(2024, 1));